pyproject.toml
[tool.poetry]
name = "big-data"
version = "0.1.0"
description = "Taller 1 de Big Data de MINE, sistema de procesamiento en tiempo real utilizando Kafka y Dask"
authors = ["Johanna Soler"]
license = "MIT"
readme = "README.md"
packages = []  # ← Agrega esta línea aquí

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"
dask = {extras = ["dataframe"], version = "^2025.2.0"}
pandas = "^2.1.4"
matplotlib = "^3.8.0"
requests = "^2.31.0"
pyarrow = "^15.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"